+++ packages/discovery-react-components/src/components/DocumentPreview/components/PdfViewerHighlight/utils/textLayout/index.ts
@@ -1,3 +1,0 @@
-export { HtmlBboxTextLayout } from './HtmlBboxTextLayout';
-export { PdfTextContentTextLayout } from './PdfTextContentTextLayout';
-export { TextMappingsTextLayout } from './TextMappingsTextLayout';